requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
GDAL==3.6.3
google-cloud-storage==2.13.0